Adding multi-selection capability
Adding features from `birch-artifact-grid` to allow us to no longer rely on that component.
This allows for the `multiple` attribute to be passed in, which will allow checkboxes to appear and multi-selection to occur.

Nothing changes as far as what is returned, it is still an array of objects.
Also updated the `demo/index.html` page to allow for usage of multi-select.